+++ examples/simple.c
@@ -60,8 +60,8 @@ int main(void) {
     return csize;
   }
 
-  printf("Compression: %zu -> %d (%.1fx)\n",
-         isize, csize, (1. * (double)isize) / csize);
+  printf("Compression: %lld -> %d (%.1fx)\n",
+         (int64_t)isize, csize, (1. * (double)isize) / csize);
 
   ret = blosc_getitem(data_out, 5, 5, data_subset);
   if (ret < 0) {
